requirements.txt
# 核心依赖
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# 阿里百炼平台
dashscope>=1.14.0

# 数据处理和分析
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0

# 深度学习框架
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0

# 数据可视化
seaborn>=0.12.0
plotly>=5.15.0

# 配置管理
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0

# 日志和监控
loguru>=0.7.0

# 测试框架
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 工具类
tqdm>=4.65.0
requests>=2.31.0